+++ chain/index.js
@@ -1,5 +1,11 @@
 const getMaxFundAmount = require('./get_max_fund_amount');
 const getNetwork = require('./get_network');
 const getTransactionRecord = require('./get_transaction_record');
+const getTransitRefund = require('./get_transit_refund');
 
-module.exports = {getMaxFundAmount, getNetwork, getTransactionRecord};
+module.exports = {
+  getMaxFundAmount,
+  getNetwork,
+  getTransactionRecord,
+  getTransitRefund,
+};