+++ peers/index.js
@@ -1,6 +1,15 @@
+const acceptsChannelOpen = require('./accepts_channel_open');
 const connectPeer = require('./connect_peer');
 const findKey = require('./find_key');
 const getLiquidity = require('./get_liquidity');
 const getPeerLiquidity =  require('./get_peer_liquidity');
+const waitForPendingOpen = require('./wait_for_pending_open');
 
-module.exports = {connectPeer, findKey, getLiquidity, getPeerLiquidity};
+module.exports = {
+  acceptsChannelOpen,
+  connectPeer,
+  findKey,
+  getLiquidity,
+  getPeerLiquidity,
+  waitForPendingOpen,
+};